// webui/api/models/RestBean_UserRoleCode.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RestBean_UserRoleCode = {
    code?: number;
    msg?: string;
    data?: RestBean_UserRoleCode.data;
};
export namespace RestBean_UserRoleCode {
    export enum data {
        SUPER_ADMIN = 'SUPER_ADMIN',
        ADMIN = 'ADMIN',
        USER = 'USER',
    }
}